+++ src/index.ts
@@ -12,3 +12,4 @@ export * from "./components/ContentWrapper.js";
 export * from "./components/Header.js";
 export * from "./components/HumanDateTime.js";
 export * from "./components/Notice.js";
+export * from "./components/Pagination.js";